+++ kraken-java-sdk/kraken-java-sdk-controller/src/main/java/com/consoleconnect/kraken/operator/controller/handler/ClientDeploymentEventHandler.java
@@ -1,7 +1,6 @@
 package com.consoleconnect.kraken.operator.controller.handler;
 
 import com.consoleconnect.kraken.operator.controller.service.ProductDeploymentService;
-import com.consoleconnect.kraken.operator.controller.tools.DeploymentErrorHelper;
 import com.consoleconnect.kraken.operator.core.client.ClientEvent;
 import com.consoleconnect.kraken.operator.core.client.ClientEventTypeEnum;
 import com.consoleconnect.kraken.operator.core.client.ClientInstanceDeployment;
@@ -11,6 +10,7 @@ import com.consoleconnect.kraken.operator.core.enums.ClientReportTypeEnum;
 import com.consoleconnect.kraken.operator.core.model.HttpResponse;
 import com.consoleconnect.kraken.operator.core.repo.EnvironmentClientRepository;
 import com.consoleconnect.kraken.operator.core.repo.UnifiedAssetRepository;
+import com.consoleconnect.kraken.operator.core.toolkit.DeploymentErrorHelper;
 import com.consoleconnect.kraken.operator.core.toolkit.JsonToolkit;
 import java.time.ZonedDateTime;
 import lombok.AllArgsConstructor;
